implemented the mapping idea for button maxi
